Add three simple, well used, ES cli commands
We add the `check` command, which displays health, problem indices,
and a quick 5 second document count to show that logs are flowing.

We also add a simple `health` command to show the over health of
the cluster and the state of the individual nodes.

Finally, we add the `indices` command, which dumps the list of
indices using MBs as the size unit.